Solve day 6, part 2
This was a bit more interesting - we overflowed u32, so had to switch to
u64. Thank goodness that was enough.

The brute-force apporach to recursively moving ahead by 10 days at a
time broke down here. I suspect because the number of fish was getting
rather silly. So I switched to considering a shoal of fish, where I only
tracked the number of each age.

Inverting the look-up tables was a pain. However, this was sufficient to
prevent the previous exponential behaviour.