Solve day 12 part 1
I was originally hoping to be able to build up a graph of actual nodes
with pointers to each other. However, that proved difficult due to the
graph not being tree-like. Also, giving the nodes mutable internal state
was causing all manner of problems.

So I went for the current approach - an immutable graph with 'pointers'
via string keys in a hashmap. It's much less safe. However, recursing to
find all possible paths is easy.